feat: fixes and moves all tests in core to new structure
