+++ packages/table-core/tests/implementation/features/row-pinning/rowSelectionFeature.test.ts
@@ -4,11 +4,13 @@ import {
   coreFeatures,
   createColumnHelper,
   rowSelectionFeature,
-} from '../src'
-import * as RowSelectionUtils from '../src/features/row-selection/rowSelectionFeature.utils'
-import { makeData } from './makeTestData'
-import type { Person } from './makeTestData'
-import type { ColumnDef } from '../src'
+} from '../../../../src'
+import * as RowSelectionUtils from '../../../../src/features/row-selection/rowSelectionFeature.utils'
+import { makeData } from '../../../makeTestData'
+import type { Person } from '../../../makeTestData'
+import type { ColumnDef } from '../../../../src'
+
+// TODO: bring up to new test structure
 
 type personKeys = keyof Person
 type PersonColumn = ColumnDef<any, Person, any>
