+++ packages/table-core/tests/performance/features/column-grouping/columnGroupingFeature.test.ts
@@ -5,11 +5,11 @@ import {
   constructTable,
   coreFeatures,
   createGroupedRowModel,
-} from '../src'
-import { createColumnHelper } from '../src/helpers/columnHelper'
-import { makeData } from './makeTestData'
-import type { Person } from './makeTestData'
-import type { ColumnDef } from '../src'
+} from '../../../../src'
+import { createColumnHelper } from '../../../../src/helpers/columnHelper'
+import { makeData } from '../../../makeTestData'
+import type { Person } from '../../../makeTestData'
+import type { ColumnDef } from '../../../../src'
 
 type personKeys = keyof Person
 type PersonColumn = ColumnDef<any, Person, any>
