+++ packages/table-core/tests/unit/core/cells/constructCell.test.ts
@@ -1,9 +1,9 @@
 import { describe, expect, it } from 'vitest'
-import { constructCell } from './constructCell'
-import { coreCellsFeature } from './coreCellsFeature'
-import type { Row } from '../../types/Row'
-import type { Column } from '../../types/Column'
-import type { Table } from '../../types/Table'
+import { constructCell } from '../../../../src/core/cells/constructCell'
+import { coreCellsFeature } from '../../../../src/core/cells/coreCellsFeature'
+import type { Row } from '../../../../src/types/Row'
+import type { Column } from '../../../../src/types/Column'
+import type { Table } from '../../../../src/types/Table'
 
 describe('constructCell', () => {
   it('should populate the cell with all core cell APIs and properties', () => {
